BAFK-341: PR feedback
