+++ aem-component-generator/src/main/java/com/bounteous/aem/compgenerator/exceptions/GeneratorException.java
@@ -26,5 +26,4 @@ public class GeneratorException extends RuntimeException {
     public GeneratorException(String errorMessage) {
         super(errorMessage);
     }
-
 }
