+++ aem-component-generator/src/main/java/com/bounteous/aem/compgenerator/javacodemodel/InterfaceBuilder.java
@@ -6,7 +6,13 @@ import com.bounteous.aem.compgenerator.models.GenerationConfig;
 import com.bounteous.aem.compgenerator.models.Property;
 import com.bounteous.aem.compgenerator.utils.CommonUtils;
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.sun.codemodel.*;
+import com.sun.codemodel.JClassAlreadyExistsException;
+import com.sun.codemodel.JCodeModel;
+import com.sun.codemodel.JDefinedClass;
+import com.sun.codemodel.JDocComment;
+import com.sun.codemodel.JMethod;
+import com.sun.codemodel.JPackage;
+import com.sun.codemodel.JType;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.text.CaseUtils;
 import org.apache.logging.log4j.LogManager;
