+++ aem-component-generator/src/main/java/com/bounteous/aem/compgenerator/javacodemodel/JavaCodeBuilder.java
@@ -1,24 +1,18 @@
 package com.bounteous.aem.compgenerator.javacodemodel;
 
-import com.bounteous.aem.compgenerator.Constants;
 import com.bounteous.aem.compgenerator.models.GenerationConfig;
 import com.bounteous.aem.compgenerator.models.Property;
-import com.sun.codemodel.CodeWriter;
-import com.sun.codemodel.JClass;
 import com.sun.codemodel.JCodeModel;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Objects;
+import java.util.Set;
 import org.apache.commons.lang3.StringUtils;
 
-import java.io.File;
-import java.io.IOException;
-import java.util.*;
 import java.util.stream.Collectors;
 
-import com.bounteous.aem.compgenerator.utils.CommonUtils;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
 public abstract class JavaCodeBuilder {
-    private static final Logger LOG = LogManager.getLogger(JavaCodeBuilder.class);
 
     protected final JCodeModel codeModel;
     protected final GenerationConfig generationConfig;
