+++ aem-component-generator/src/main/java/com/bounteous/aem/compgenerator/javacodemodel/RenameFileCodeWriter.java
@@ -19,7 +19,6 @@ public class RenameFileCodeWriter extends FileCodeWriter {
     @Override
     protected File getFile(JPackage pkg, String fileName) throws IOException {
         File dir;
-
         if (pkg.isUnnamed()) {
             dir = target;
         } else {
@@ -31,9 +30,7 @@ public class RenameFileCodeWriter extends FileCodeWriter {
         }
 
         String filePath = new File(dir, fileName).getPath();
-
         File codeFile = CommonUtils.getNewFileAtPathAndRenameExisting(filePath);
-
         return codeFile;
     }
 }
