+++ aem-component-generator/src/main/java/com/bounteous/aem/compgenerator/models/Options.java
@@ -22,8 +22,11 @@ package com.bounteous.aem.compgenerator.models;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.List;
+import org.apache.commons.lang3.ArrayUtils;
 
 public class Options implements BaseModel {
+
+    private final static String[] DEFAULT_ADAPTABLES = new String[]{"request"};
 
     @JsonProperty("generic-javadoc")
     private boolean hasGenericJavadoc;
@@ -42,6 +45,9 @@ public class Options implements BaseModel {
 
     @JsonProperty("content-exporter")
     private boolean allowExporting;
+
+    @JsonProperty("model-adaptables")
+    private String[] modelAdaptables;
 
     @JsonProperty("properties")
     private List<Property> properties;
@@ -98,6 +104,17 @@ public class Options implements BaseModel {
 
     public void setAllowExporting(boolean allowExporting) {
         this.allowExporting = allowExporting;
+    }
+
+    public String[] getModelAdaptables() {
+        if (ArrayUtils.isEmpty(modelAdaptables)) {
+            return DEFAULT_ADAPTABLES;
+        }
+        return modelAdaptables;
+    }
+
+    public void setModelAdaptables(final String[] modelAdaptables) {
+        this.modelAdaptables = modelAdaptables;
     }
 
     public List<Property> getProperties() {
